refactor: ディレクトリ構成の変更、typeを別ファイルに切り出し、不要なimportの削除
