+++ src/types/chartType.ts
@@ -1,0 +1,5 @@
+// 各軸の設定値
+export type selectValues = {
+  x: string;
+  y: string;
+};
