+++ src/types/selector.ts
@@ -1,0 +1,10 @@
+export type Option = {
+  value: string;
+  label: string;
+};
+
+export type SelectProps = {
+  options: Option[];
+  defaultValue?: string | undefined;
+  onChange?: (value: string) => void;
+};
